Better codegen quality when emitting degenerate switches.
Related to dotnet/coreclr#7914
Where the offending pattern was observed in the IL generated for the Kestrel web server and was recommended as an easy improvement.

Becasue of cascaded dispatching into "try" regions , it is not uncommon to see async state machine to contain degenerate switches like

IL_0008:  switch     (
                          IL_003b,
                          IL_003b,
                          IL_003b,
                          IL_003b,
                          IL_003b,
                          IL_003b,
                          IL_003b,
                          IL_0729)

Note numerous cases all leading to the same target.
We can trivially emit such switches as just a range-check (that switch would need to do anyways), without any actual "switching".
Since all cases lead to the same label anyways, there is no need to switch once we know the value is in range of the switch.

Fixes:#14878